Backend/requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bert-score==0.3.13
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
datasets==4.8.4
dill==0.4.1
evaluate==0.4.6
faiss-cpu==1.13.2
filelock==3.28.0
Flask==3.1.3
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
googleapis-common-protos==1.74.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
pip==23.1.2
propcache==0.4.1
protobuf==5.29.6
proto-plus==1.27.2
pyarrow==23.0.1
pyasn1_modules==0.4.2
pyasn1==0.6.3
pycparser==3.0
pydantic_core==2.46.2
pydantic==2.13.2
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rouge_score==0.1.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==65.5.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.7
xxhash==3.6.0
yarl==1.23.0
